move to v1 package, start cleaning up whitespace
