+++ fabric-tool-attributes-v1/src/main/java/net/fabricmc/fabric/api/tools/v1/ToolAttributeHolder.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.api.tools;
+package net.fabricmc.fabric.api.tools.v1;
 
 import com.google.common.collect.ImmutableSetMultimap;
 import com.google.common.collect.Multimap;
