+++ fabric-tool-attributes-v1/src/main/java/net/fabricmc/fabric/impl/tools/EntityToolActor.java
@@ -1,6 +1,6 @@
 package net.fabricmc.fabric.impl.tools;
 
-import net.fabricmc.fabric.api.tools.ToolActor;
+import net.fabricmc.fabric.api.tools.v1.ToolActor;
 import net.minecraft.entity.LivingEntity;
 
 public class EntityToolActor implements ToolActor<LivingEntity> {
